added comments, moved searchBar.jsx directory location
