+++ frontend/src/Dashboard/AdminEmployeesTable/AdminEmployeesTable.jsx
@@ -1,18 +1,15 @@
 import React, { useState, useEffect } from 'react'
 import { Table } from 'flowbite-react'
 import EmployeesListNew from './EmployeesListNew'
-import SearchBar from '../SearchBar'
+import SearchBar from './SearchBar'
 
-export default function AdminEmployeesTable({
-    employees,
-    setEmployees,
-    openModal,
-}) {
-
+export default function AdminEmployeesTable({ employees, openModal }) {
     const [filteredEmployees, setFilteredEmployees] = useState(null)
-    
-    useEffect(() => {setFilteredEmployees(employees)}, [employees])
 
+    // Create a copy of employees array to mutate leaving original employees array untouched
+    useEffect(() => {
+        setFilteredEmployees(employees)
+    }, [employees])
 
     return (
         <>
@@ -20,7 +17,10 @@ export default function AdminEmployeesTable({
                 <Table hoverable>
                     <Table.Head>
                         <Table.HeadCell colSpan={6}>
-                            <SearchBar employees={employees} setFilteredEmployees={setFilteredEmployees}/>
+                            <SearchBar
+                                employees={employees}
+                                setFilteredEmployees={setFilteredEmployees}
+                            />
                         </Table.HeadCell>
                     </Table.Head>
                     <Table.Head className="sticky top-0 bg-white z-10">
