+++ frontend/src/Dashboard/AdminEmployeesTable/SearchBar.jsx
@@ -8,6 +8,7 @@ export default function SearchBar({ employees, setFilteredEmployees }) {
         return searchResults
     }
 
+    // Run of the mill text field change handler with the addition of a simple search function executed on each letter typed
     const handleChange = (event) => {
         const { value } = event.target
         setSearchString(value)
